Make HTTP a behaviour module
